+++ packages/djangoflow_analytics/djangoflow_mixpanel_analytics/test/helpers/mock_mixpanel_analytics.dart
@@ -1,0 +1,4 @@
+import 'package:mixpanel_flutter/mixpanel_flutter.dart';
+import 'package:mockito/annotations.dart';
+@GenerateNiceMocks([MockSpec<Mixpanel>()])
+export 'mock_mixpanel_analytics.mocks.dart';