+++ packages/djangoflow_analytics/djangoflow_mixpanel_analytics/test/src/features/user_properties_removal/facebook_user_property_remover_test.dart
@@ -1,0 +1,33 @@
+import 'package:djangoflow_mixpanel_analytics/djangoflow_mixpanel_analytics.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
+
+import '../../../helpers/helpers.dart';
+
+void main() {
+  late MockMixpanel mockMixpanelAnalytics;
+  late MixpanelUserPropertyRemover userPropertyRemover;
+  const kEmailKey = 'email';
+
+  setUp(() {
+    mockMixpanelAnalytics = MockMixpanel();
+    userPropertyRemover = MixpanelUserPropertyRemover(mockMixpanelAnalytics);
+  });
+
+  test('canHandle() should return true for MixpanelRemovableUserProperty', () {
+    final event = MixpanelRemovableUserProperty(kEmailKey);
+    expect(userPropertyRemover.canHandle(event), true);
+  });
+
+  test('perform() should call setUserId with null value', () {
+    final event = MixpanelRemovableUserProperty(kEmailKey);
+    when(
+      mockMixpanelAnalytics.unregisterSuperProperty(any),
+    ).thenAnswer((_) {});
+
+    userPropertyRemover.perform(event);
+    verify(
+      mockMixpanelAnalytics.unregisterSuperProperty(event.key),
+    ).called(1);
+  });
+}