+++ packages/djangoflow_analytics/djangoflow_analytics/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - 23 Jan 2023
+
+- Added `performers` getter for `DjangoflowAnalytics`.
+- Added unit tests
+
 ## [0.0.1] - 18 Jan 2023
 
 - Initialization of the library through the init method on the DjangoflowAnalytics class