+++ packages/djangoflow_deep_link/djangoflow_deep_link_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.1.1] - 23 Jan 2023
+
+- Fixed typo `Intial` to `Initial` in places
+
+## [1.1.0] - 23 Jan 2023
+
+- Initialize listening to deep link stream (`getLinkStream`) immediately
+- Added unit tests
+
 ## [1.0.1] - 13 Jan 2023
 
 - Fix exports