+++ packages/djangoflow_error_reporters/djangoflow_error_reporter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.2] - Jan 23 2023
+
+- Added unit tests
+
 ## [0.0.1] - Jan 16 2023
 
 - Provides basic functionality for initializing, reporting and updating user information