+++ packages/djangoflow_push_notification/djangoflow_fcm/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.3] - 23 Jan 2023
+
+- Fixed returning `authorizationStatus` with token
+
+## [0.0.2] - 23 Jan 2023
+
+- Added unit tests
+- Fixed typo `Intial` to `Initial` in places
+- Updated dependencies
+
 ## [0.0.1] - 15 Jan
 
 - Implemented DjangoflowFCMBloc, Repository and BlocListeners(for Token and RemoteMessage updates)